Add files via upload
In this project, I downloaded data from the Ghana Stock Exchange where Ghana Oil Company was listed. I used Jupyter Notebook to run Machine Learning Algorithms to predict the stock price for the next year. The Goal of this Project was to use Monte Carlo Simulation to Predict the Closing Stock Price of Ghana Oil Company Ltd (GOIL).

A presentation slide was prepared and presented.